// libs/core/src/gen/reliable_pb.ts
// @generated by protoc-gen-es v2.10.2 with parameter "target=ts"
// @generated from file reliable.proto (package reliable, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file reliable.proto.
 */
export const file_reliable: GenFile = /*@__PURE__*/
  fileDesc("Cg5yZWxpYWJsZS5wcm90bxIIcmVsaWFibGUiUgoGUmVsTXNnEhEKCXN0cmVhbV9pZBgBIAEoBBILCgNzZXEYAiABKAQSDwoHcGF5bG9hZBgDIAEoDBIXCg9yZXN5bmNfcmVxdWlyZWQYBCABKAgiMgoLUmVsRGVsaXZlcnkSFAoMcHVibGlzaGVyX2lkGAEgASgJEg0KBWZyYW1lGAIgASgMIkQKB1JlbE5hY2sSEQoJc3RyZWFtX2lkGAEgASgEEhAKCGZyb21fc2VxGAIgASgEEhQKDHB1Ymxpc2hlcl9pZBgDIAEoCSI2CgpSZWxDb250cm9sEiEKBG5hY2sYASABKAsyES5yZWxpYWJsZS5SZWxOYWNrSABCBQoDbXNnYgZwcm90bzM");

/**
 * Publisher -> SFU -> Subscriber (ordered/reliable channel framing)
 *
 * @generated from message reliable.RelMsg
 */
export type RelMsg = Message<"reliable.RelMsg"> & {
  /**
   * @generated from field: uint64 stream_id = 1;
   */
  streamId: bigint;

  /**
   * @generated from field: uint64 seq = 2;
   */
  seq: bigint;

  /**
   * @generated from field: bytes payload = 3;
   */
  payload: Uint8Array;

  /**
   * @generated from field: bool resync_required = 4;
   */
  resyncRequired: boolean;
};

/**
 * Describes the message reliable.RelMsg.
 * Use `create(RelMsgSchema)` to create a new message.
 */
export const RelMsgSchema: GenMessage<RelMsg> = /*@__PURE__*/
  messageDesc(file_reliable, 0);

/**
 * SFU wraps RelMsg before delivering to subscriber
 *
 * @generated from message reliable.RelDelivery
 */
export type RelDelivery = Message<"reliable.RelDelivery"> & {
  /**
   * @generated from field: string publisher_id = 1;
   */
  publisherId: string;

  /**
   * encoded RelMsg
   *
   * @generated from field: bytes frame = 2;
   */
  frame: Uint8Array;
};

/**
 * Describes the message reliable.RelDelivery.
 * Use `create(RelDeliverySchema)` to create a new message.
 */
export const RelDeliverySchema: GenMessage<RelDelivery> = /*@__PURE__*/
  messageDesc(file_reliable, 1);

/**
 * Subscriber -> Publisher (retransmit request)
 *
 * @generated from message reliable.RelNack
 */
export type RelNack = Message<"reliable.RelNack"> & {
  /**
   * @generated from field: uint64 stream_id = 1;
   */
  streamId: bigint;

  /**
   * @generated from field: uint64 from_seq = 2;
   */
  fromSeq: bigint;

  /**
   * @generated from field: string publisher_id = 3;
   */
  publisherId: string;
};

/**
 * Describes the message reliable.RelNack.
 * Use `create(RelNackSchema)` to create a new message.
 */
export const RelNackSchema: GenMessage<RelNack> = /*@__PURE__*/
  messageDesc(file_reliable, 2);

/**
 * @generated from message reliable.RelControl
 */
export type RelControl = Message<"reliable.RelControl"> & {
  /**
   * @generated from oneof reliable.RelControl.msg
   */
  msg: {
    /**
     * @generated from field: reliable.RelNack nack = 1;
     */
    value: RelNack;
    case: "nack";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message reliable.RelControl.
 * Use `create(RelControlSchema)` to create a new message.
 */
export const RelControlSchema: GenMessage<RelControl> = /*@__PURE__*/
  messageDesc(file_reliable, 3);